pyproject.toml
[project]
name = "pubsub-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-pubsub>=2.35.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml",
    "tabulate",
]
